Add incron task to update the default group in the console on cert changes
This commit adds a new class to the fundamentals module and includes it by
default. The class adds an incron (file watcher) task to update the default
group in the dashboard whenever a new cert is added. This is so that new
machines are classified automatically instead of waiting for a two minute
interval.